fix: SetRepositoryPolicyCommand differs by ECR type
