+++ src/ecr.ts
@@ -8,7 +8,6 @@ import {
   Repository,
 } from '@aws-sdk/client-ecr'
 import { promises as fs } from 'fs'
-import { setRepositoryPolicy } from './common'
 
 type Inputs = {
   repository: string
@@ -92,4 +91,12 @@ const putLifecyclePolicy = async (client: ECRClient, repositoryName: string, pat
 
   await client.send(new PutLifecyclePolicyCommand({ repositoryName, lifecyclePolicyText }))
   core.info(`successfully put lifecycle policy ${path} to repository ${repositoryName}`)
+}
+
+export const setRepositoryPolicy = async (client: ECRClient, repositoryName: string, path: string): Promise<void> => {
+  const policyText = await fs.readFile(path, { encoding: 'utf-8' })
+  core.debug(`setting the repository policy ${path} to repository ${repositoryName}`)
+
+  await client.send(new SetRepositoryPolicyCommand({ repositoryName, policyText }))
+  core.info(`successfully set repository policy ${path} to repository ${repositoryName}`)
 }
