+++ src/ecr_public.ts
@@ -3,9 +3,10 @@ import {
   ECRPUBLICClient,
   DescribeRepositoriesCommand,
   CreateRepositoryCommand,
+  SetRepositoryPolicyCommand,
   Repository,
 } from '@aws-sdk/client-ecr-public'
-import { setRepositoryPolicy } from './common'
+import { promises as fs } from 'fs'
 
 type Inputs = {
   repository: string
@@ -78,3 +79,11 @@ const isRepositoryNotFoundException = (e: unknown) => e instanceof Error && e.na
 
 // ECR Public does not support the lifecycle policy
 // https://github.com/aws/containers-roadmap/issues/1268
+
+export const setRepositoryPolicy = async (client: ECRPUBLICClient, repositoryName: string, path: string): Promise<void> => {
+  const policyText = await fs.readFile(path, { encoding: 'utf-8' })
+  core.debug(`setting the repository policy ${path} to repository ${repositoryName}`)
+
+  await client.send(new SetRepositoryPolicyCommand({ repositoryName, policyText }))
+  core.info(`successfully set repository policy ${path} to repository ${repositoryName}`)
+}
